Don't set Content-Type header for HTML codec.
Setting Content-Type to x-www-form-urlencoded, even for a GET request, causes
the TeamCity REST API to fail with a 500 error. Empirically it works better
for the HTML codec just to delete the Content-Type header. We haven't yet
engaged this with a REST API that requires form-coded POST data anyway.

While we're at it, dict.update({key: value}) is just a clumsy way of saying
dict[key] = value. Change to the more direct, readable form.