Introduce RESTService operations' basepath= parameter.
This feature lets a RESTService instance consumer disregard whatever partial
pathname was constructed into the RESTService and construct a new path from
the beginning. For instance, it's convenient to construct a RESTService for
the TeamCity REST API with the partial path 'httpAuth/app/rest', since every
REST API path starts with that prefix. But previously, that left a consumer of
that RESTService no way to query a path _without_ that prefix. basepath
provides that functionality, e.g. to reach the Linden plugin.

This means adding optional basepath=None to each existing operation, and
enhancing RESTService._url() to accept both basepath and path.