Convert map keys to UnicodeType before serializing rather than str.
The trouble with converting keys to literally str() is that if, in Python 2,
the key value is unicode to start with, converting to str() means attempting
to encode it as 8-bit bytes using the 'ascii' codec -- which fails. Convert to
UnicodeType instead, which is an alias for unicode in Python 2 and str in 3.