Add RESTService(basepath=) param, also clone() method.
When a RESTService consumer explicitly instantiates that RESTService, passing
a URL with not only the hostname but also a partial path is convenient. Every
subsequent use of that RESTService need only append the part of the path that
diverges between uses.

The llbuildutils.services module was introduced when our TeamCity hostname
changed (with its migration to EC2). At that point, every script that
explicitly instantiated a TeamCity RESTService had to be changed to address
the new TeamCity location. We realized that it was desirable to centralize
knowledge, not only of the hostname but of the entire conventional RESTService
configuration needed to reach the target service.

But that posed a problem. The RESTServices delivered by llbuildutils.services
typically do contain a baseurl with a partial path. That partial path is set
by a module we explicitly want to be able to change with changes to the target
service. Can a given consumer script count on the partial path set by that
module? Should it?

Since llbuildutils.services.get_service() et al. provide a mechanism to
override any RESTService constructor parameter, the approach we choose is to
add a basepath= constructor parameter. A simple consumer can make assumptions
about the partial path delivered by llbuildutils.services; a robust consumer
can explicitly override the basepath for the requested instance, while still
letting the services module provide the target hostname.

Separately, add a RESTService.clone() method. Sometimes it's important to be
able to parallelize queries to the same service, e.g. with eventlet. But you
can't reuse the same RESTService instance for concurrent queries because you
can't reuse its contained requests.Session instance. In one consumer script,
we built a "RESTService factory" to capture RESTService configuration
parameters so we could generate RESTService instances with the same
configuration as needed. But clone() is both simpler and more robust, since
it's lexically adjacent to RESTService.__init__(): a person maintaining either
should recognize the need to keep the other in sync.

clone() works better than a RESTService factory with the llbuildutils.services
usage pattern: such a factory must store configuration parameters that
otherwise belong to llbuildutils.services. With clone(), a consumer can simply
request an instance and then clone() it as needed for parallelism.

RESTService didn't literally store every one of its constructor parameters
before, using session_params and proxy_hostport "on the fly." For clone(), we
now store both as passed so we can pass the same argument values to a
subsequent constructor call.

Also fix a RESTError bug: if any of its variable parameters were passed as
unicode, the format() call would break.