Add future.types.newstr as a supported string type if future is installed.
see docs: https://python-future.org/what_else.html#str

This is to support code that has been ported to python3 but wants to maintain
backwards compatibility with python2 using the future package.  Note that we
do not want to require a dependency on future.  If future is not installed,
then we proceed with out it.